Separate out swift-native sub-packages
Separate out the swift-native into distinct subpackages that mirror the target
packges. This avoids conflicts when building Swift tools for the host system,
by avoiding development-only artefacts being installed in the native sysroot.

It should also serve to draw attention to new and/or removed files in future
Swift versions so the target recipes can be updated.